+++ src/Client/IparapheurV5/Model/Premis.php
@@ -24,6 +24,9 @@ final class Premis
     {
         $intellectual = new PremisObject();
         $intellectual->type = PremisObject::INTELLECTUAL_ENTITY;
+        if (empty($fileToSign->dossierTitre)) {
+            throw new RuntimeException('Le titre du dossier (dossierTitre) est obligatoire');
+        }
         $intellectual->originalName = $fileToSign->dossierTitre;
 
         $significantProperties = [];